[#8] 권한 관리 기능 구현

+++ src/main/java/hello/hellospring/service/PostService.java
@@ -1,6 +1,8 @@
 package hello.hellospring.service;
 
 
+import hello.hellospring.common.exception.UnAuthorizationException;
+import hello.hellospring.domain.Member;
 import hello.hellospring.domain.Post;
 import hello.hellospring.dto.PostDto;
 import hello.hellospring.repository.PostRepository;
@@ -10,7 +12,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
@@ -18,16 +19,17 @@ import java.util.NoSuchElementException;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
-    private final MemberService memberService;
+    private final LoginService loginService;
 
-
-    public void saveNewPost(PostDto.Request request) {
-        Post post = request.toEntity(memberService.findMemberById(request.getMemberId()));
+    public void saveNewPost(PostDto.Request request, Member member) {
+        Post post = request.toEntity(member);
         postRepository.save(post);
     }
 
     public void updatePost(Post post, PostDto.Request request) {
-        post.update(request);
+        if (isOwner(post)) {
+            post.update(request);
+        }
     }
 
     @Transactional(readOnly = true)
@@ -38,11 +40,22 @@ public class PostService {
     }
 
     public void deletePost(Long id) {
-        postRepository.deleteById(id);
+        Post post = postRepository.findById(id).orElseThrow(NoSuchElementException::new);
+        if (isOwner(post)) {
+            postRepository.deleteById(id);
+        }
     }
 
     @Transactional(readOnly = true)
     public Page<Post> findAllPosts(Pageable pageable) {
         return postRepository.findAll(pageable);
+    }
+
+    public boolean isOwner(Post post) {
+        Member member = loginService.getLoggedInMember();
+        if (member != post.getMember()) {
+            throw new UnAuthorizationException();
+        }
+        return true;
     }
 }
